convert subrepo perms filtering into a job
